Extract datagrid's enqueue functionality into a re-usable service; put it to use in EuiIcon
